Fixed a lot of issues: "arduino.h" should be written with capital A, made recv structures actually work with VESC, used stricter typing

+++ local_datatypes.h
@@ -10,7 +10,7 @@ struct bldcMeasure {
 	float avgId;
 	float avgIq;
 	float dutyNow;
-	long rpm;
+	float rpm;
 	float inpVoltage;
 	float ampHours;
 	float ampHoursCharged;
